backend/requirements.txt
Django==5.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
qrcode[pil]==7.4.2
Pillow==10.3.0
django-cors-headers==4.3.1
mysqlclient==2.2.4